Documentation for SintPowerCase.jl

This is a package intended for reading in power system data from various formats and storing it in dataframes. The dataframes should be compatible with the MATPOWER format.

Method documentation

SintPowerCase.get_bus_row — Method

Returns the row number of a bus given by id

SintPowerCase.get_gen_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_id_idx — Method

Returns the row number of an element given by id

SintPowerCase.get_incidence_matrix — Method

get_incidence_matrix(network::PowerGraphBase)::Array{Float64}
Returns the susceptance vector for performing a dc power flow.

SintPowerCase.get_incidence_matrix — Method

get_incidence_matrix(network::PowerGraphBase)::Array{Float64}
Returns the susceptance vector for performing a dc power flow.

SintPowerCase.get_load_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_n_buses — Method

Returns the number of buses in the case.

SintPowerCase.get_n_lines — Method

Returns the number of lines in the case.

SintPowerCase.get_susceptance_vector — Method

get_susceptance_vector(network::PowerGraphBase)::Array{Float64}
Returns the susceptance vector for performing a dc power flow.

SintPowerCase.put_back_line! — Method

Sets branch to out of service

SintPowerCase.take_out_line! — Method

Sets branch to out of service

SintPowerCase.to_ppc — Method

Convert the case to the PYPOWER Case format.