Documentation for SintPowerCase.jl

This is a package intended for reading in power system data from various formats and storing it in dataframes. The dataframes should be compatible with the MATPOWER format.

Method documentation

SintPowerCase.get_all_buses_aggregated_power — Method

Aggregates the power for generators or loads for all buses

Inputs: df: Dataframe with either the power for all generators or loads.

SintPowerCase.get_bus_row — Method

Returns the row number of a bus given by id

SintPowerCase.get_dc_admittance_matrix — Method

get_dc_admittance_matrix(case)
Returns the admittance matrix for performing a dc power flow.
Inputs:
    case: The power system data.
    consider_status: If branch in service status should be considered.

SintPowerCase.get_gen_buses_power — Method

Returns the power for generator buses for os

SintPowerCase.get_gen_buses_power — Method

Returns the sum of production at generator buses.

SintPowerCase.get_gen_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_gen_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_id_idx — Method

Returns the row number of an element given by id

SintPowerCase.get_incidence_matrix — Method

get_incidence_matrix(case)::Array{Float64}
Returns the incidence matrix of a power system.

Inputs:
    case: The power system to get the susceptance vector for.
    consider_status: Whether or not branch in service status should be
    considered.

SintPowerCase.get_load_buses_power — Method

Returns the power demamd for load buses for os

SintPowerCase.get_load_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_load_indices — Method

Return indices of the buses with loads.

SintPowerCase.get_n_buses — Method

Returns the number of buses in the case.

SintPowerCase.get_n_lines — Method

Returns the number of lines in the case.

SintPowerCase.get_power_injection_vector — Method

Returns the power injection vector.

SintPowerCase.get_susceptance_vector — Method

get_susceptance_vector(case)
Returns the susceptance vector for performing a dc power flow.

Inputs:
    case: The power system to get the susceptance vector for.
    consider_status: Whether or not branch in service status should be
        considered.

SintPowerCase.is_load_bus — Method

is_load_bus(case, id::String)
Returns true if the bus bus_id is a load.

SintPowerCase.put_back_line! — Method

Sets branch to out of service

SintPowerCase.take_out_line! — Method

Sets branch to out of service

SintPowerCase.to_ppc — Method

Convert the case to the PYPOWER Case format.